Complete refactoring of GeoGigFeatureReader
Uses the backgrounding feature reader, and is
built using FeatureReaderBuilder.

Uses the spatial index when available.

Next step is to support full filtering directly
from NodeRefs and avoid fetching RevFeatures when
possible (i.e. the filter is fully supported and
all the required properties by the Query are available
in the index).

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>